requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
